server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pdfplumber==0.10.3
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.5.1
google-generativeai==0.3.2
tiktoken==0.6.0 